Updated game title and enhanced level generation
- Rebranded game title from "Dusk Moon" to "Codename Nightfall" for a fresh identity.
- Cleaned up code formatting for readability, including consistent spacing around operators and corrected indentation.
- Improved depth calculation in the Mob entity for more precise underwater behavior.
- Enhanced the level generator to maintain a better balance of terrain features, ensuring a minimum presence of vital elements and avoiding excessive stairs generation.
- Reinstated previously commented-out code in LevelGen to diversify the environment with the addition of flowers and cacti.

+++ src/main/java/de/cirrus/dusk/entities/Mob.java
@@ -105,8 +105,8 @@ public class Mob extends Entity {
                 // the water. 1.0f -- max depth, 0.0f -- we're on the land
                 // (swimming == false).
                 // TODO: implement depth map instead
-                depth += getNextTile() != Tile.water && depth > 0f ? -Math.sqrt(depth) * 0.05f : getPreviousTile() != Tile.water && depth < 1f ? Math
-                        .sqrt(1 - depth) * 0.05f : 0f;
+                depth += (float) (getNextTile() != Tile.water && depth > 0f ? -Math.sqrt(depth) * 0.05f : getPreviousTile() != Tile.water && depth < 1f ? Math
+                                        .sqrt(1 - depth) * 0.05f : 0f);
             } else {
                 swimming = false;
                 depth = 0f;
